requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.15.0
pandas==2.1.3
numpy==1.24.3
paho-mqtt==1.6.1
streamlit==1.28.1
pydantic==2.5.0
python-multipart==0.0.6
plotly==5.18.0
requests==2.31.0